requirements.txt
asyncio==3.4.3
dotenv==0.9.9
fastapi==0.116.1
httpx==0.28.1
langchain==0.3.27
langchain-core==0.3.72
langchain-pymupdf4llm==0.4.1
langchain-text-splitters==0.3.9
llama-cloud-services==0.6.53
pydantic==2.11.7
python-dotenv==1.1.1
transformers==4.49.0
typing==3.7.4.3
uvicorn==0.35.0
psutil
groq
rank_bm25
sentence-transformers
scikit-learn
llama-index
PyMuPDF
numpy
Pillow==10.4.0
pandas 
easyocr
timm
unstructured
